Move object-related API tests
